Fix issue setting xhr.timeout to an object
Previously, if the request timeout was Nothing, it would set the timeout
on the XHR request to the Elm representation of Maybe, i.e.
`{ ctor: 'Nothing' }`. This was causing issues in Elm Native, where
Objective-C was type-checking the value of timeout as a number, and it
was coming back as a dictionary.

This solves the issue by only setting the timeout on the XHR request if
the value is `Just`, and setting it to the value itself instead of the
Elm representation of that value.