Switch to simplified API
Lets people handle bytes and files, but overall is simpler than what we
had before.

So it should be easier for pretty much everybody, but the root
motivation was that progress cannot be tracked with a subscription
unless the request is initiated from a command. That meant that the
configuration for Cmd and Task could not be the same, thus inspiring
the idea to just get rid of the Request type all together.

This closes the door on having a (Request a) type in the core library
that can have a parallel2 function. A bunch of time has passed since
that initial design, and I think that sort of thing requires separate
design considerations, particularly around tracking progress. So
instead of trying to fit it all into one thing right now, the idea is
to focus on providing the features that actually exist today.

(Note: Parallel requests can currently be done with simultaneous
commands. It requires some modeling, but considering the full range of
special cases, it is unclear at this time if an API can cover that
scenario better in a way that is both general and pleasant.)